add tests

+++ lib/data/models/local_chv_malaria_case_model.dart
@@ -17,7 +17,7 @@ class LocalChvMalariaCaseModel implements IdentifiableWithUuidObject {
   @override
   String? name;
 
-  int? age;
+  double? age;
 
   @override
   DateTime? created;
