+++ lib/data/models/remote_chv_malaria_case_model.dart
@@ -17,7 +17,7 @@ class RemoteChvMalariaCaseModel implements IdentifiableWithUuidObject {
   @override
   String? name;
 
-  int? age;
+  double? age;
 
   @override
   DateTime? created;
